requirements.txt
dvc
scikit-learn
pandas
numpy
apache-airflow
yellowbrick
mlflow
seaborn